Applied db changes

+++ Micro.Fin/Fin.ApplicationCore/Entities/Loan.cs
@@ -16,7 +16,7 @@ namespace Fin.ApplicationCore.Entities
         public decimal InitialLoanAmount { get; set; }
         public DateTime DateGranted { get; set; }
         public int PeriodMonths { get; set; }
-        public int Interest { get; set; }
+        public decimal Interest { get; set; }
         public string Security { get; set; }
         public decimal PropertyValue { get; set; }
         public decimal CapitalOutstanding { get; set; }
